# apps/app-4/docs/openapi.yaml
openapi: 3.0.0
info:
  title: MediSync Telemedicine API
  version: 2.0.0
  description: |
    API completa para gestão de teleconsultas, agendamentos, prontuários eletrônicos,
    receitas, atestados, pagamentos e muito mais para clínicas e hospitais.
    
    ## Autenticação
    A API usa JWT (JSON Web Token) para autenticação. Inclua o token no header:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Endpoints públicos: 10 requisições/minuto
    - Endpoints autenticados: 100 requisições/minuto
    
    ## Códigos de Erro
    - 400: Dados inválidos
    - 401: Não autenticado
    - 403: Sem permissão
    - 404: Recurso não encontrado
    - 409: Conflito (ex: horário já agendado)
    - 429: Muitas requisições
    - 500: Erro interno
  contact:
    name: Suporte MediSync
    email: suporte@medisync.com.br
  license:
    name: Proprietário
servers:
  - url: http://localhost:8080
    description: Desenvolvimento
  - url: https://api.medisync.com.br
    description: Produção
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login de usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas
  /auth/register:
    post:
      tags: [Auth]
      summary: Registro de novo paciente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuário criado
...
components:
  schemas:
...
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
        acceptTerms:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [ADMIN, MEDICO, PACIENTE]
        phone:
          type: string
        specialty:
          type: string
        crm:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido no login